requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.10.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
dataclasses-json==0.6.6
deepdiff==7.0.1
Deprecated==1.2.14
deprecation==2.1.0
effdet==0.4.1
emoji==2.12.1
filelock==3.14.0
filetype==1.2.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.5.0
functions-framework==3.7.0
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-cloud-vision==3.7.2
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.2
humanfriendly==10.0
idna==3.7
intel-openmp==2021.4.0
iopath==0.1.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.2.2
langchain-community==0.2.2
langchain-core==0.2.4
langchain-text-splitters==0.2.1
langdetect==1.0.9
langsmith==0.1.71
layoutparser==0.3.4
lxml==5.2.2
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.9.0
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
omegaconf==2.3.0
onnx==1.16.1
onnxruntime==1.18.0
opencv-python==4.10.0.82
ordered-set==4.1.0
orjson==3.10.3
packaging==23.2
pandas==2.2.2
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.0
pikepdf==9.0.0
pillow==10.3.0
pillow_heif==0.16.0
portalocker==2.8.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.7
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pyparsing==3.1.2
pypdf==4.2.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pyreadline3==3.4.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
pywin32==306
PyYAML==6.0.1
rapidfuzz==3.9.3
regex==2024.5.15
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
safetensors==0.4.3
scipy==1.13.1
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
sympy==1.12.1
tabulate==0.9.0
tbb==2021.12.0
tenacity==8.3.0
timm==1.0.3
tokenizers==0.19.1
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
transformers==4.41.2
typing-inspect==0.9.0
typing_extensions==4.12.1
tzdata==2024.1
unstructured==0.14.4
unstructured-client==0.23.0
unstructured-inference==0.7.33
unstructured.pytesseract==0.3.12
urllib3==2.2.1
watchdog==4.0.1
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
